CS-524 Limit size of picture uploads on Together
Images for projects on Together were already resized.
I've just changed the compression function to uses the pillow algorithm instead of an homemade one
I've also remove an feature in this function that was not really useful and understandable which was using the file size as an indicator instead of the bitmap size.