Assign call task to event ambassador if church rep
From the communication job, retrieves the linked records if and only if they're recurring contracts, iterate over them to find an eventual related event through origin.

If an event is found, get the ambassador and find the related employee and its job. If he's a church rep, assign him the call task, otherwise assign it to the default user (previous behavior)